+++ server/.gitignore
@@ -122,3 +122,5 @@ dist
 .pnp.*
 
 # End of https://www.toptal.com/developers/gitignore/api/node
+
+package-lock.json
